Moved suppress_warnings to test-util file
Moved suppress_warnigns from core.util to core.tests.util, as it should only really be used in testcases anyways.

Removed unused imports.
