+++ activity_calendar/tests/test_views.py
@@ -17,7 +17,7 @@ from activity_calendar.views import CreateSlotView, ActivityMomentWithSlotsView,
 from activity_calendar.forms import *
 
 from core.models import ExtendedUser as User
-from core.util import suppress_warnings
+from core.tests.util import suppress_warnings
 
 from . import mock_now
